requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
supabase==2.3.4
openai==1.12.0
python-dotenv==1.0.0
chromadb==0.4.24
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
textblob==0.17.1
transformers==4.36.2
torch==2.1.2
pydantic==2.5.3
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
jwt==1.3.1
cryptography>=41.0.0
requests==2.31.0 